requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1  # Added for environment variables